Fix inconsistencies in handling of LLBPosixFileDetails
For directory entries, the LLBPosixFileDetails are stored in both the
LLBCASFileTree and the corresponding LLBDirectoryEntry, and this change
fixes a few places where they were not kept consistent.

Also, the LLBPosixFileDetails are stored as an optional field of
LLBDirectoryEntry, where a nil value indicates that the entry has the
default value (which depends on the file type). If the default value is
specified explicitly but then normalized to nil, there were a few places
where the optional field needs to be cleared.